+++ htmlresources/objectbrowser/js/objectbrowser.js
@@ -270,7 +270,8 @@ class ObjectBrowser {
                     me.execObjCommand(key, $(this)[0], kind);
                 },
                 items: {
-                    "definition": {name: "Go to definition"},
+                    "definition": {name: "Go to definition (project file or server definition)"},
+                    "localdefinition": {name: "Open definition from app file"},
                     "shownewtab": {name: "Open symbol in new Tab"},
                     "sep1": "---------",
                     "runinwebclient": {