+++ src/langserver/toolsGetLibrarySymbolLocationResponse.ts
@@ -1,0 +1,5 @@
+import { ALSymbolSourceLocation } from "../symbollibraries/alSymbolSourceLocation";
+
+export class ToolsGetLibrarySymbolLocationResponse {
+    location: ALSymbolSourceLocation | undefined;
+}