+++ src/symbollibraries/alSymbolSourceLocation.ts
@@ -1,0 +1,7 @@
+import { TextRange } from "./textRange";
+
+export class ALSymbolSourceLocation {
+    schema: string | undefined;
+    sourcePath: string | undefined;
+    range: TextRange | undefined;
+}